Move production_listing template tag to production_tags
