+++ awards/templates/awards/_recommendation_summary_listing.html
@@ -1,4 +1,4 @@
-{% extends "tags/production_listing.html" %}
+{% extends "productions/tags/production_listing.html" %}
 
 {% block columns %}
     {{ block.super }}
