+++ awards/templates/awards/_recommendation_summary_listing_row.html
@@ -1,4 +1,4 @@
-{% extends "tags/production_listing_row.html" %}
+{% extends "productions/tags/production_listing_row.html" %}
 
 {% block cells %}
     {{ block.super }}
