+++ awards/templates/awards/candidates.html
@@ -1,6 +1,6 @@
 {% extends 'base.html' %}
 
-{% load demoscene_tags laces %}
+{% load production_tags laces %}
 
 {% block html_title %}{{ event.name }} - {{ category.name }} candidates - Demozoo{% endblock %}
