+++ awards/templates/awards/report.html
@@ -1,6 +1,6 @@
 {% extends 'base.html' %}
 
-{% load awards_tags demoscene_tags %}
+{% load awards_tags production_tags %}
 
 {% block html_title %}{{ event.name }} - {{ category.name }} recommendations - Demozoo{% endblock %}
