+++ bbs/templates/bbs/show.html
@@ -1,7 +1,5 @@
 {% extends "base.html" %}
-{% load demoscene_tags releaser_tags compress %}
-{% load safe_markdown %}
-
+{% load demoscene_tags production_tags releaser_tags compress safe_markdown %}
 
 {% block html_title %}{{ bbs.name }} - Demozoo{% endblock %}
