+++ maintenance/templates/maintenance/duplicate_external_links.html
@@ -1,5 +1,5 @@
 {% extends "maintenance/base.html" %}
-{% load compress demoscene_tags %}
+{% load production_tags %}
 
 {% block html_title %}Duplicate external links - Demozoo{% endblock %}
