+++ maintenance/templates/maintenance/randomised_production_report.html
@@ -1,5 +1,5 @@
 {% extends "maintenance/base.html" %}
-{% load compress demoscene_tags %}
+{% load production_tags %}
 
 {% block html_title %}{{ title }} - Demozoo{% endblock %}
