+++ parties/templates/parties/show.html
@@ -1,7 +1,5 @@
 {% extends "base.html" %}
-{% load demoscene_tags compress %}
-{% load safe_markdown %}
-
+{% load demoscene_tags production_tags compress safe_markdown %}
 
 {% block html_title %}{{ party.name }} - Demozoo{% endblock %}
