+++ platforms/templates/platforms/show.html
@@ -1,6 +1,5 @@
 {% extends "base.html" %}
-{% load safe_markdown compress demoscene_tags %}
-
+{% load production_tags safe_markdown %}
 
 {% block html_title %}{{ platform.name }} - Demozoo{% endblock %}
