+++ productions/templates/music/index.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
-{% load compress demoscene_tags laces %}
+{% load production_tags laces %}
 
 
 {% block html_title %}Music - Demozoo{% endblock %}
