+++ productions/templates/productions/tagged.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
-{% load safe_markdown compress demoscene_tags laces %}
+{% load safe_markdown production_tags laces %}
 
 {% block html_title %}Productions tagged '{{ tag.name }}' - Demozoo{% endblock %}
