+++ productions/templates/productions/tags/production_listing.html
@@ -11,7 +11,7 @@
     <tbody>
         {% block rows %}
             {% for production, screenshot in productions_and_screenshots %}
-                {% include "tags/production_listing_row.html" %}
+                {% include "productions/tags/production_listing_row.html" %}
             {% endfor %}
         {% endblock %}
     </tbody>
