+++ productions/templatetags/production_tags.py
@@ -1,4 +1,7 @@
 from django import template
+from django.conf import settings
+
+from productions.models import Screenshot
 
 
 register = template.Library()
@@ -44,3 +47,20 @@ def production_core_details(context, production, subpage=False):
         )
 
     return ctx
+
+
+@register.inclusion_tag("productions/tags/production_listing.html")
+def production_listing(productions, show_screenshots=False, show_prod_types=False, mark_excludable=False):
+    if show_screenshots:
+        screenshots = Screenshot.select_for_production_ids([prod.id for prod in productions])
+    else:
+        screenshots = {}
+
+    productions_and_screenshots = [(production, screenshots.get(production.id)) for production in productions]
+    return {
+        "productions_and_screenshots": productions_and_screenshots,
+        "show_screenshots": show_screenshots,
+        "show_prod_types": show_prod_types,
+        "mark_excludable": mark_excludable,
+        "site_is_writeable": settings.SITE_IS_WRITEABLE,
+    }
